Sleech version 1.1.0
Added the slag description to both the (S)Leech and Breath of Terra grenade mods to highlight that both grenade mods are dual-elemental. Also adjusted the damage of the Breath's initial explosion to set it en par with a purple Tediore grenade to give it a bit more Oomph and add some fun to the mix.